Move get_circuit_assignments to separate mod
The vamp_ir_utils module consists of definitions that will eventually be
incorporated into the vamp-ir library
