+++ taiga_halo2/src/circuit/mod.rs
@@ -9,4 +9,5 @@ pub mod curve;
 pub mod hash_to_curve;
 pub mod note_encryption_circuit;
 pub mod schnorr_signature;
+mod vamp_ir_utils;
 pub mod vp_examples;
